levers: the MoE forward and MTP drafter paths, and a seam for model_levers.rs
`MtpHead::forward_one` runs once per DRAFTED TOKEN and read
`ATLAS_MTP_DEBUG_NORMS` FOUR separate times inside one call — every one of
them only to decide whether to do nothing. `MoeFfnLayer::forward` and its
batched-gate and K=2 siblings read four more per layer per decode token.

`fp32_routing_active` is the one that needed real plumbing. It was read from
SIX call sites on the decode path, and the env test is the LAST term of a
five-way conjunction whose other four terms are properties of the layer's own
weights and kernels. So the lever moved to `ModelLevers` and the
preconditions stayed on the layer: the function now takes `&ModelLevers`,
threaded through `FfnLayer::fp32_routing_active` and its four callers.
Passing it rather than reading it is the point — a site that forgets it
fails to build.

`model_levers.rs` reached 483 lines, so the resolution table and the three
constructors move to `model_levers_resolve.rs`. The seam is deliberate: the
parent file is now the DECLARATION, one documented field per lever, which is
what a reader asking "what does this flag do" wants; the child is the single
place that touches the environment, which is what a reader asking "how is it
spelled" wants. `resolve` is PRIVATE — `ModelLevers::get` is the only way
out of it — and the tests reach in as its sibling to drive `from_values`
directly, so they still exercise the production resolver rather than a copy.

The guard table grows from ten files to fifteen, and it now guards
`model_levers_resolve.rs` itself with only `from_env` allowed: a read
anywhere else in that file is a SECOND resolution path, which is the thing
this module exists to prevent.

★ A CONTROL I REFUSED TO COUNT. The first negative control for the extended
table printed `test result: ok`. It had not run: reinjecting the raw read
left the `levers` parameter unused, the build failed, and a build failure
prints no verdict at all. A guard that "passes" because nothing executed is
exactly the failure the control exists to detect. Redone with a read that
compiles, it goes red naming `layers/moe/forward_k2.rs:99 in forward_k2`, and
green again on restore.

Verified: `cargo test --workspace --locked` 77/77, clippy 0 warnings, all
four lever files under the 500-line cap.
