+++ crates/spark-model/src/layers/qwen3_ssm/trait_decode.rs
@@ -97,7 +97,7 @@ impl Qwen3SsmLayer {
         // ATLAS_FP32_ROUTING: emit the MoE-input norm in FP32 (router_in) so the
         // gate GEMM runs at full precision — removes the bf16-store rounding that
         // flips experts on gfx1151. bf16 normed2 + residual are unchanged.
-        if self.ffn.fp32_routing_active() && self.residual_add_rms_norm_gatef32_k.0 != 0 {
+        if self.ffn.fp32_routing_active(ctx.levers) && self.residual_add_rms_norm_gatef32_k.0 != 0 {
             ops::residual_add_rms_norm_gatef32(
                 ctx.gpu,
                 self.residual_add_rms_norm_gatef32_k,
